+++ app/src/main/java/woowacourse/shopping/ui/MainViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.di.DIContainer
-import com.example.di.FieldInject
+import com.example.di.annotations.FieldInject
 import com.example.di.annotations.Qualifier
 import kotlinx.coroutines.launch
 import woowacourse.shopping.data.repository.CartRepository
