+++ di/src/main/java/com/example/di/annotations/FieldInject.kt
@@ -1,4 +1,4 @@
-package com.example.di
+package com.example.di.annotations
 
 @Target(AnnotationTarget.FIELD)
 annotation class FieldInject
