+++ di/src/main/java/com/example/di/annotations/Qualifier.kt
@@ -1,6 +1,6 @@
 package com.example.di.annotations
 
-@Target(AnnotationTarget.FIELD)
+@Target(AnnotationTarget.FIELD, AnnotationTarget.CONSTRUCTOR)
 @Retention(AnnotationRetention.RUNTIME)
 annotation class Qualifier(
     val value: String,
